[FEAT]생각했던 것 리스트 조회 개발

+++ src/main/java/com/async/main/domain/cleaningGauge/dto/CleaningGaugeDto.java
@@ -17,7 +17,7 @@ public class CleaningGaugeDto {
     private Integer idx;
 
     @JsonProperty("thought_idx")
-    private Long thoughtIdx;
+    private Integer thoughtIdx;
 
     @JsonProperty("space_type")
     private String spaceType;
